+++ packages/growerp_core/lib/src/domains/authenticate/authenticate.dart
@@ -1,4 +1,3 @@
 export 'blocs/blocs.dart';
 export 'models/models.dart';
 export 'views/views.dart';
-export 'widgets/widgets.dart';