+++ packages/growerp_core/lib/src/domains/common/views/home_form.dart
@@ -16,7 +16,7 @@ import '../../domains.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:responsive_framework/responsive_framework.dart';
-import '../../../templates/@templates.dart';
+import '../../../templates/templates.dart';
 import 'package:global_configuration/global_configuration.dart';
 
 class HomeForm extends StatefulWidget {