+++ packages/growerp_core/lib/src/domains/users/views/user_dialog.dart
@@ -20,7 +20,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:image_picker/image_picker.dart';
-import '../../../templates/@templates.dart';
+import '../../../templates/templates.dart';
 import 'package:responsive_framework/responsive_wrapper.dart';
 import '../../domains.dart';
 import '../../../api_repository.dart';